+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version 3.0.1 *(In Development)*
 * Fix: `onOptionsItemSelected()` not being called in fragments if the activity
   version returns `false`.
 * Fix: `onCreateOptionsMenu()` not being called in fragments on Android 3.0+.
+* New: Enable action item text display on pre-Android 3.0 by calling
+  `requestWindowFeature` with `Window.FEATURE_ENABLE_ACTION_BAR_WATSON_TEXT`.
 
 
 Version 3.0.0 *(2010-06-05)*